Add CurlProperty component and update CurlProperties
