+++ components/layout/execute/index.ts
@@ -1,4 +1,5 @@
 export * from "./CurlCommand";
+export * from "./CurlProperty";
 export * from "./CurlProperties";
 export * from "./CurlPropertyInput";
 export * from "./ExecutePanel";
